requirements.txt
elevenlabs>=1.0.0
python-dotenv
pydub
ffmpeg-python
